requirements.txt
pytest
pytest-html
allure-pytest
allure-python-commons
selenium
requests
dacite
webdriver-manager
